st-backend/requirements.txt
fastapi
uvicorn[standard]
motor
pymongo
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]
pydantic-settings
python-dotenv
bcrypt
scikit-learn
joblib
numpy
pandas
